Move ui actions to seperate files
