+++ src/app/oidc/getConfirmInteraction.js
@@ -1,0 +1,9 @@
+const oidc = require('./oidcServer');
+
+const getConfirmInteraction = async (req, res) => {
+  await oidc.interactionFinished(req, res, {
+    consent: {},
+  });
+};
+
+module.exports = getConfirmInteraction;
