+++ src/app/oidc/getDevUsernamePassword.js
@@ -1,0 +1,5 @@
+const getDevUsernamePassword = (req, res) => {
+  res.render('oidc/views/usernamepassword', { uuid: req.params.uuid });
+};
+
+module.exports = getDevUsernamePassword;
