+++ src/app/oidc/oidcServer.js
@@ -1,0 +1,48 @@
+const uuid = require('uuid/v4');
+const config = require('./../../infrastructure/Config');
+// const logger = require('./../../infrastructure/logger');
+const Provider = require('oidc-provider');
+const Accounts = require('./../../infrastructure/Accounts');
+const logoutAction = require('./../logout');
+
+const oidc = new Provider(`${config.hostingEnvironment.protocol}://${config.hostingEnvironment.host}:${config.hostingEnvironment.port}`, {
+  clientCacheDuration: 60,
+  logoutSource: logoutAction,
+  findById: Accounts.findById,
+  claims: {
+    // scope: [claims] format
+    openid: ['sub'],
+    email: ['email'],
+    profile: ['email', 'given_name', 'family_name'],
+  },
+  interactionUrl(ctx) {
+    return `/interaction/${ctx.oidc.uuid}`;
+  },
+  async interactionCheck(ctx) {
+    if (!ctx.oidc.session.sidFor(ctx.oidc.client.clientId)) {
+      const sid = uuid();
+      ctx.oidc.session.sidFor(ctx.oidc.client.clientId, sid);
+      await ctx.oidc.session.save();
+    }
+
+    return false;
+  },
+  // TODO deployment configuration
+  features: {
+    // disable the packaged interactions
+    devInteractions: false,
+    claimsParameter: true,
+    clientCredentials: true,
+    discovery: true,
+    encryption: true,
+    introspection: true,
+    registration: true,
+    request: true,
+    requestUri: true,
+    revocation: true,
+    rejectUnauthorized: false,
+    sessionManagement: true,
+  },
+});
+
+module.exports = oidc;
